PrintUtils divided into classes with only one purpose: MessageParser (parse strings with String.format), MessagePrinter (print message on console and record it in log) and ScannerLogger (to record in log and pass user inputs to the app)
